Renamed strategy directories to filter
