+++ pathetic-api/src/main/java/org/patheloper/api/pathing/filter/PathValidationContext.java
@@ -1,4 +1,4 @@
-package org.patheloper.api.pathing.strategy;
+package org.patheloper.api.pathing.filter;
 
 import lombok.Value;
 import org.patheloper.api.snapshot.SnapshotManager;
