+++ pathetic-api/src/main/java/org/patheloper/api/pathing/filter/filters/WaterPathFilter.java
@@ -1,9 +1,9 @@
-package org.patheloper.api.pathing.strategy.strategies;
+package org.patheloper.api.pathing.filter.filters;
 
 import lombok.NonNull;
 import org.bukkit.Material;
-import org.patheloper.api.pathing.strategy.PathValidationContext;
-import org.patheloper.api.pathing.strategy.PathFilter;
+import org.patheloper.api.pathing.filter.PathValidationContext;
+import org.patheloper.api.pathing.filter.PathFilter;
 import org.patheloper.api.snapshot.SnapshotManager;
 import org.patheloper.api.wrapper.PathPosition;
